refactor: move InstrumentIcon into own directory
