+++ apps/web/src/features/instruments/components/InstrumentCard/InstrumentCard.tsx
@@ -4,7 +4,7 @@ import { Card } from '@douglasneuroinformatics/libui/components';
 import type { UnilingualInstrumentSummary } from '@opendatacapture/schemas/instrument';
 import { useTranslation } from 'react-i18next';
 
-import { InstrumentIcon } from './InstrumentIcon';
+import { InstrumentIcon } from '../InstrumentIcon';
 
 export type InstrumentCardProps = {
   instrument: UnilingualInstrumentSummary;
