+++ apps/web/src/features/instruments/components/InstrumentIcon/InstrumentIcon.stories.tsx
@@ -1,0 +1,9 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { InstrumentIcon } from './InstrumentIcon';
+
+type Story = StoryObj<typeof InstrumentIcon>;
+
+export default { component: InstrumentIcon } as Meta<typeof InstrumentIcon>;
+
+export const Default: Story = {};
