+++ apps/web/src/features/instruments/components/InstrumentIcon/InstrumentIcon.tsx
