+++ apps/web/src/features/instruments/components/InstrumentIcon/index.ts
@@ -1,0 +1,1 @@
+export * from './InstrumentIcon';
